# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WireGuard Mesh Manager API",
    "description": "Backend API for managing WireGuard VPN networks and devices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/export": {
      "post": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Export Networks From Request",
        "description": "Export networks from a request body.",
        "operationId": "export_networks_from_request_api_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/export/download": {
      "post": {
        "tags": [
          "export",
          "export"
        ],
        "summary": "Download Export",
        "description": "Download export data as a ZIP file.",
        "operationId": "download_export_api_export_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit/events": {
      "get": {
        "tags": [
          "audit",
          "audit"
        ],
        "summary": "List Audit Events",
        "description": "List audit events with pagination and filtering.",
        "operationId": "list_audit_events_api_audit_events_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by network ID",
              "title": "Network Id"
            },
            "description": "Filter by network ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (ISO format)",
              "title": "Start Date"
            },
            "description": "Filter by start date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (ISO format)",
              "title": "End Date"
            },
            "description": "Filter by end date (ISO format)"
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor name",
              "title": "Actor"
            },
            "description": "Filter by actor name"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resource type",
              "title": "Resource Type"
            },
            "description": "Filter by resource type"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (starts from 1)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (starts from 1)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of events per page (max 1000)",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Number of events per page (max 1000)"
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include event details in response",
              "default": false,
              "title": "Include Details"
            },
            "description": "Include event details in response"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Audit Events Api Audit Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit/export": {
      "get": {
        "tags": [
          "audit",
          "audit"
        ],
        "summary": "Export Audit Events",
        "description": "Export audit events with optional filtering.",
        "operationId": "export_audit_events_api_audit_export_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by network ID",
              "title": "Network Id"
            },
            "description": "Filter by network ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (ISO format)",
              "title": "Start Date"
            },
            "description": "Filter by start date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (ISO format)",
              "title": "End Date"
            },
            "description": "Filter by end date (ISO format)"
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor name",
              "title": "Actor"
            },
            "description": "Filter by actor name"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format (json or csv)",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format (json or csv)"
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include event details",
              "default": true,
              "title": "Include Details"
            },
            "description": "Include event details"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Export Audit Events Api Audit Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/export/download": {
      "get": {
        "tags": [
          "audit",
          "audit"
        ],
        "summary": "Download Audit Events",
        "description": "Download audit events as a file.",
        "operationId": "download_audit_events_api_audit_export_download_get",
        "parameters": [
          {
            "name": "network_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by network ID",
              "title": "Network Id"
            },
            "description": "Filter by network ID"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by start date (ISO format)",
              "title": "Start Date"
            },
            "description": "Filter by start date (ISO format)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by end date (ISO format)",
              "title": "End Date"
            },
            "description": "Filter by end date (ISO format)"
          },
          {
            "name": "actor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by actor name",
              "title": "Actor"
            },
            "description": "Filter by actor name"
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by action type",
              "title": "Action"
            },
            "description": "Filter by action type"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format (json or csv)",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format (json or csv)"
          },
          {
            "name": "include_details",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include event details",
              "default": true,
              "title": "Include Details"
            },
            "description": "Include event details"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportData": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/ExportMetadata"
          },
          "networks": {
            "items": {
              "$ref": "#/components/schemas/WireGuardNetworkExport"
            },
            "type": "array",
            "title": "Networks",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "metadata"
        ],
        "title": "ExportData",
        "description": "Complete export data structure.",
        "example": {
          "metadata": {
            "description": "Backup of production networks",
            "exported_at": "2024-01-01T00:00:00Z",
            "exported_by": "admin@example.com",
            "version": "1.0"
          },
          "networks": []
        }
      },
      "ExportMetadata": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version",
            "default": "1.0"
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          },
          "exported_by": {
            "type": "string",
            "title": "Exported By"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "exported_at",
          "exported_by"
        ],
        "title": "ExportMetadata",
        "description": "Metadata for export files."
      },
      "ExportRequest": {
        "properties": {
          "network_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network Ids"
          },
          "include_configs": {
            "type": "boolean",
            "title": "Include Configs",
            "default": true
          },
          "include_api_keys": {
            "type": "boolean",
            "title": "Include Api Keys",
            "default": false
          },
          "format": {
            "type": "string",
            "enum": [
              "json",
              "zip"
            ],
            "title": "Format",
            "default": "json"
          }
        },
        "type": "object",
        "title": "ExportRequest",
        "description": "Request schema for exporting networks."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WireGuardNetworkExport": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "network_cidr": {
            "type": "string",
            "pattern": "^\\d+\\.\\d+\\.\\d+\\.\\d+/\\d+$",
            "title": "Network Cidr"
          },
          "dns_servers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dns Servers"
          },
          "mtu": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mtu"
          },
          "persistent_keepalive": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 86400.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Persistent Keepalive"
          },
          "private_key_encrypted": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Key Encrypted"
          },
          "public_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 56,
                "minLength": 44
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Key"
          },
          "preshared_key_encrypted": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Preshared Key Encrypted"
          },
          "interface_properties": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface Properties",
            "description": "Additional interface properties"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationExport"
            },
            "type": "array",
            "title": "Locations",
            "default": []
          },
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceExport"
            },
            "type": "array",
            "title": "Devices",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "network_cidr"
        ],
        "title": "WireGuardNetworkExport",
        "description": "Complete network data for export.\n\nFor mesh topology, networks do not have their own WireGuard keys.\nKey fields are optional to support both mesh (no keys) and legacy formats."
      },
...
    }
  }
}